requirements.txt
tensorflow
flask
requests
sqlalchemy
transformers
flask_wtf
sentencepiece
sacremoses
bcrypt
flask-SQLAlchemy
flask-Bcrypt
flask-Login
SQLAlchemy
WTForms
Jinja2
flask-migrate
python-dotenv
google-generativeai